backend/requirements.txt
# Facebook Marketing API
facebook-business==24.0.1

# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11

# Data Processing
pandas==2.3.3

# Configuration
python-dotenv==1.2.1

# FastAPI & Web Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.0.0,<3.0.0

# Google Sheets API
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0

# Excel Export (alternative)
openpyxl==3.1.5
